+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented here. The format is based
 on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixes
+
+- only add default forbid policy to fields that do not have a (custom) policy yet
+
 ## 2023-07-26 0.3.0
 
 ### Features